add ButtonConnect component & refactor
