+++ src/pages/Bridge/ActionPanel/BridgeActionPanel.tsx
@@ -2,11 +2,6 @@ import React, { useCallback } from 'react'
 import { ChainId } from '@swapr/sdk'
 import { ButtonPrimary } from '../../../components/Button'
 import { useNetworkSwitch } from '../../../hooks/useNetworkSwitch'
-import {
-  useModalOpen,
-  useNetworkSwitcherPopoverToggle,
-  useWalletSwitcherPopoverToggle
-} from '../../../state/application/hooks'
 import { BridgeStep } from '../utils'
 import { NetworkSwitcher } from './NetworkSwitcher'
 import { BridgeButton } from './BridgeButton'
@@ -19,8 +14,7 @@ import ProgressSteps from '../../../components/ProgressSteps'
 import Column from '../../../components/Column'
 import { useBridgeActionPanel } from './useBridgeActionPanel'
 import { ApprovalState } from '../../../hooks/useApproveCallback'
-import { ApplicationModal } from '../../../state/application/actions'
-import { useUnsupportedChainIdError } from '../../../hooks'
+import { ButtonConnect } from '../../../components/ButtonConnect'
 
 export type BridgeActionPanelProps = {
   account: string | null | undefined
@@ -43,8 +37,6 @@ export const BridgeActionPanel = ({
   isNetworkConnected
 }: BridgeActionPanelProps) => {
   const { selectNetwork } = useNetworkSwitch()
-  const toggleWalletSwitcherPopover = useWalletSwitcherPopoverToggle()
-  const toggleNetworkSwitcherPopover = useNetworkSwitcherPopoverToggle()
 
   const {
     approvalState,
@@ -55,9 +47,6 @@ export const BridgeActionPanel = ({
     isArbitrum,
     hasAmount
   } = useBridgeActionPanel()
-  const networkSwitcherPopoverOpen = useModalOpen(ApplicationModal.NETWORK_SWITCHER)
-  const unsupportedChainIdError = useUnsupportedChainIdError()
-  const isSwitchNetwork = networkSwitcherPopoverOpen || unsupportedChainIdError
 
   const handleSelectFromNetwork = useCallback(() => {
     selectNetwork(fromNetworkChainId)
@@ -70,14 +59,7 @@ export const BridgeActionPanel = ({
   const selectPanel = () => {
     // No wallet
     if (!account) {
-      return (
-        <ButtonPrimary
-          onClick={isSwitchNetwork ? toggleNetworkSwitcherPopover : toggleWalletSwitcherPopover}
-          disabled={networkSwitcherPopoverOpen}
-        >
-          {isSwitchNetwork ? 'Switch network' : 'Connect wallet'}
-        </ButtonPrimary>
-      )
+      return <ButtonConnect />
     }
 
     // Change network
