+++ src/pages/Swap/index.tsx
@@ -16,16 +16,11 @@ import { ArrowWrapper, SwapCallbackError, SwitchTokensAmountsContainer, Wrapper 
 import TradePrice from '../../components/swap/TradePrice'
 import TokenWarningModal from '../../components/TokenWarningModal'
 import ProgressSteps from '../../components/ProgressSteps'
-import { useActiveWeb3React, useUnsupportedChainIdError } from '../../hooks'
+import { useActiveWeb3React } from '../../hooks'
 import { useAllTokens, useCurrency } from '../../hooks/Tokens'
 import { ApprovalState, useApproveCallbackFromTrade } from '../../hooks/useApproveCallback'
 import { useSwapCallback } from '../../hooks/useSwapCallback'
 import useWrapCallback, { WrapType } from '../../hooks/useWrapCallback'
-import {
-  useModalOpen,
-  useNetworkSwitcherPopoverToggle,
-  useWalletSwitcherPopoverToggle
-} from '../../state/application/hooks'
 import { Field } from '../../state/swap/actions'
 import {
   useDefaultsFromURLSearch,
@@ -43,7 +38,6 @@ import Loader from '../../components/Loader'
 import { useTargetedChainIdFromUrl } from '../../hooks/useTargetedChainIdFromUrl'
 import { ROUTABLE_PLATFORM_LOGO } from '../../constants'
 import QuestionHelper from '../../components/QuestionHelper'
-import { ApplicationModal } from '../../state/application/actions'
 
 // Landing Page Imports
 import './../../theme/landingPageTheme/stylesheet.css'
@@ -55,6 +49,7 @@ import CommunityLinks from './../../components/LandingPageComponents/CommunityLi
 import BlogNavigation from './../../components/LandingPageComponents/BlogNavigation'
 import Hero from './../../components/LandingPageComponents/layout/Hero'
 import Footer from './../../components/LandingPageComponents/layout/Footer'
+import { ButtonConnect } from '../../components/ButtonConnect'
 
 const RotatedRepeat = styled(Repeat)`
   transform: rotate(90deg);
@@ -114,9 +109,6 @@ export default function Swap() {
 
   const { account, chainId } = useActiveWeb3React()
   const theme = useContext(ThemeContext)
-
-  // toggle wallet when disconnected
-  const toggleWalletSwitcherPopover = useWalletSwitcherPopoverToggle()
 
   // for expert mode
   const [isExpertMode] = useExpertModeManager()
@@ -291,11 +283,6 @@ export default function Swap() {
     [onCurrencySelection]
   )
 
-  const networkSwitcherPopoverOpen = useModalOpen(ApplicationModal.NETWORK_SWITCHER)
-  const unsupportedChainIdError = useUnsupportedChainIdError()
-  const isSwitchNetwork = networkSwitcherPopoverOpen || unsupportedChainIdError
-  const toggleNetworkSwitcherPopover = useNetworkSwitcherPopoverToggle()
-
   return (
     <>
       <TokenWarningModal
@@ -400,12 +387,7 @@ export default function Swap() {
                 )}
                 <div>
                   {!account ? (
-                    <ButtonPrimary
-                      onClick={isSwitchNetwork ? toggleNetworkSwitcherPopover : toggleWalletSwitcherPopover}
-                      disabled={networkSwitcherPopoverOpen}
-                    >
-                      {isSwitchNetwork ? 'Switch network' : 'Connect wallet'}
-                    </ButtonPrimary>
+                    <ButtonConnect />
                   ) : showWrap ? (
                     <ButtonPrimary disabled={Boolean(wrapInputError)} onClick={onWrap}>
                       {wrapInputError ??
